Initial Commit
If you found this so far hidden repository by being sneaky and curious, welcome! This is the home of the Unity plugin for traVRsal. I had stored it on my HD for too long and became nervous that I don't have a backup if something goes wrong. So, I decided to finally commit.

It's also in quite good shape already. You can create levels from a template. Multi-level editing is supported. Levels can be packaged for local testing (including auto-detect of changed levels) or for server deployment. Levels can already be uploaded to my S3 server from where they can already be consumed.

There is an experimental function that will create an index of a level including screenshots of all prefabs so that reuse is more easily possible. Also the initial tile-set for Tiled is already included. And there are even project settings registered that you can fill out to have TMX files automatically converted to JSON.

So now the journey begins. It's a big moment for me and I hope for many great levels to be created in the future. Let's rock!